+++ lib/presentation/pages/register/register_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:gradient_app_bar/gradient_app_bar.dart';
-import 'package:zc_mobile_app/presentation/_bloc/register/register_bloc.dart';
+import 'package:zc_mobile_app/presentation/blocs/register/register_bloc.dart';
 import 'package:zc_mobile_app/presentation/widgets/register/register_form.dart';
 import 'package:zc_mobile_app/user_repository/user_repository.dart';
