+++ lib/presentation/widgets/login/google_login_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:zc_mobile_app/presentation/_bloc/login/login.dart';
+import 'package:zc_mobile_app/presentation/blocs/login/login.dart';
 
 class GoogleLoginButton extends StatelessWidget {
   @override
